requirements.txt
absl-py==0.4.1
astor==0.7.1
astroid==2.0.2
beautifulsoup4==4.6.1
boto==2.49.0
boto3==1.9.40
botocore==1.12.40
bz2file==0.98
cachetools==2.1.0
Calculator==1.1.0
certifi==2018.4.16
chardet==3.0.4
Click==7.0
colorthief==0.2.1
colour==0.1.5
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
fake-useragent==0.1.10
future==0.16.0
gast==0.2.0
gensim==3.6.0
gif-for-cli==1.1.1
giphy-client==1.0.0
giphypop==0.3
google-api-core==1.5.0
google-auth==1.5.1
google-cloud==0.34.0
google-cloud-texttospeech==0.2.0
google-images-download==2.3.0
Google-Search-API==1.1.14
googleapis-common-protos==1.5.3
grpcio==1.14.2
h5py==2.8.0
idna==2.7
imageio==2.4.1
isort==4.3.4
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==2.6.11
matplotlib==3.0.2
mccabe==0.6.1
modules==1.0.0
moviepy==0.2.3.5
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.3.1
murmurhash==1.0.1
numpy==1.15.4
opencv-python==3.4.2.17
pandas==0.23.4
Pillow==5.2.0
plac==0.9.6
preshed==2.0.1
progressbar==2.5
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.17.1
pydub==0.23.0
pyimgur==0.6.0
pylint==2.1.1
PyMaxflow==1.2.11
pyparsing==2.3.0
pyrouge==0.1.3
python-dateutil==2.7.5
pytrends==4.4.0
pytube==9.3.3
pytz==2018.7
PyYAML==3.13
regex==2018.1.10
requests==2.20.0
rope==0.11.0
rsa==4.0
s3transfer==0.1.13
scenedetect==0.5
scipy==1.1.0
seaborn==0.9.0
selenium==2.53.6
six==1.11.0
smart-open==1.7.1
spacy==2.0.16
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
thinc==6.12.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.24.0
typed-ast==1.1.0
typing==3.6.6
ujson==1.35
Unidecode==1.0.22
urllib3==1.24
vcrpy==2.0.1
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
x256==0.0.3
yarl==1.2.0
